requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
apscheduler
python-json-logger
httpx
python-multipart
pytest
pytest-asyncio

# Core libs used across services (unversioned)
pymongo
requests
openai
langgraph
typing_extensions
annotated-types
anyio
async-timeout
cachetools
certifi
charset-normalizer
click
distro
dnspython
exceptiongroup
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
h11
html2text
httpcore
httplib2
httptools
idna
jiter
jsonpatch
jsonpointer
langchain
langchain-core
langchain-openai
langchain-text-splitters
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
oauthlib
orjson
ormsgpack
packaging
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyparsing
PyYAML
regex
requests-oauthlib
requests-toolbelt
rsa
sniffio
SQLAlchemy
starlette
tenacity
tiktoken
tqdm
typing-inspection
uritemplate
urllib3
uvicorn
uvloop
watchfiles
websockets
xxhash
zstandard

